Extract interface and type definitions
