+++ src/features/calamities/models/calamity.ts
@@ -1,0 +1,8 @@
+import type { SeverityLevel } from "./data";
+
+export interface Calamity {
+    id: number,
+    location: string;
+    severity: SeverityLevel;
+    description: string;
+  }
