+++ src/features/calamities/models/data.ts
@@ -1,0 +1,1 @@
+export type SeverityLevel = "critical" | "high" | "medium" | "low"
